General update, see description.
Faction stashes were removed a looooong time ago so the rule was removed as well. I think that line was related to faction NPC locations for their guns, but those NPCs are restricted to the faction they're setup for so the rule is no longer needed. Quite outdated. It's good to look at the rules every now and again with fresh eyes :)

Included that door rams cannot be used to break into properties by normal players. This helps fill the gap that happens occasionally where people are using the door ram to break into properties without an admin because technically it doesn't require admin supervision to just click the item and unlock the interior.